Fixed bug
A user could enter only a group_id and hit enter. The program was using
the last
value of student_number instead of throwing an error that none was
entered. Now there is a check after reading the student_number. I don’t
think this scenario will actually be tested though.